added scheduled task to delete expired tokens

+++ src/main/java/pl/coderslab/charity/web/controller/ForgotPasswordController.java
@@ -1,18 +1,17 @@
 package pl.coderslab.charity.web.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.mail.SimpleMailMessage;
-import org.springframework.security.core.parameters.P;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import pl.coderslab.charity.passwordReset.Dto.PasswordResetDto;
 import pl.coderslab.charity.passwordReset.service.PasswordResetTokenService;
 import pl.coderslab.charity.user.entity.User;
 import pl.coderslab.charity.user.service.UserService;
-import pl.coderslab.charity.web.email.EmailDetails;
 import pl.coderslab.charity.web.email.EmailService;
 
 import javax.servlet.http.HttpServletRequest;
@@ -43,8 +42,7 @@ public class ForgotPasswordController {
         String token = UUID.randomUUID().toString();
         userService.createPasswordResetTokenForUser(user, token);
 
-        String s = emailService.sendMail(emailService.constructResetTokenEmail(request.getServletPath(), token, user));
-
+        emailService.sendMail(emailService.constructResetTokenEmail(request.getServletPath(), token, user));
 
         return "/passwordReset/mailsent";
     }
@@ -63,20 +61,20 @@ public class ForgotPasswordController {
 
     @PostMapping("/reset/passwordreset")
     public String changePassword(@ModelAttribute("passwordReset") @Valid PasswordResetDto passwordResetDto) {
-       String result = passwordResetTokenService.validatePasswordResetToken(passwordResetDto.getToken());
+        String result = passwordResetTokenService.validatePasswordResetToken(passwordResetDto.getToken());
 
-//       if (result != null) {
-//           return "passwordReset/passwordreset" + "?token=" + passwordResetDto.getToken();
-//       }
+        if (result != null) {
+            return "passwordReset/passwordreset" + "?token=" + passwordResetDto.getToken();
+        }
 
-       Optional user = userService.getUserByToken(passwordResetDto.getToken());
-       if (user.isPresent()) {
-           userService.changePassword((User) user.get(), passwordResetDto.getNewPassword());
+        Optional user = userService.getUserByToken(passwordResetDto.getToken());
+        if (user.isPresent()) {
+            userService.changePassword((User) user.get(), passwordResetDto.getNewPassword());
             passwordResetTokenService.deleteToken(passwordResetDto.getToken());
-           return "passwordReset/succespasswordreset";
-       } else {
-           return "passwordReset/passwordreset";
-       }
+            return "passwordReset/succespasswordreset";
+        } else {
+            return "passwordReset/passwordreset";
+        }
 
     }
 }
